clm3_6_61
iMost of our finidat files have weights incompatible with our new fsurdat files. Hence, we went back to allowing
the weights to be different and to using the finidat weights so that answers would be the same as before.
Also hardwire the logfile for datm and clm so that can run testsuite on jaguar. Also add in cppdef required
for breeze.