clm4_5_14_r229
Update cime to a version with an updated top level directory structure. It also fixes a problem
with a CLM test.

The new cime version has a new directory structure, and updates the XML schema for namelist files
from verison 1 to version 2. Currently the perl namelist code can NOT read version 2. Hence, a local
copy of the version 1 files for the drv_flds_in fields was added to CLM. Also the use of the datm
namelist files for some tools was dropped -- I think that's OK though.

Fix two minor bugs.

The new cime has a few improvements. It now adjusts the PIO namelist when the PE layout changes.
There's a new option to case.setup "--no-adjust-pio" to not do that. --mpilib option was added
to create_test. You can remove your bld directory area without having to rerun case.setup.