Move final flux calculation into implicitEuler
This simplifies the use of implicitEuler. Martyn Clark supported doing
this, since it will be needed in general:
https://github.com/NCAR/clm-ctsm/pull/12#discussion_r137650687

Also, move the endrun clal into implicitEuler, rather than requiring the
caller to check the error code. This simplifies the interface, and works
more smoothly now that the final flux calculation is also in this
routine.