clm4_5_16_r255
Changes for crop on as well as enabling carbon isotopes when crop is on. Fix some small answer changing issues.
Also bring in new timeseries datasets for C14 as well as now C13. The C14 dataset has hardwired latitude sectors
it uses (so previous file can NOT be used). Properly handle and update C13/C14 variables when crop is on.
Handle CIsoFluxCalc calls for a list of crop specific fields. Add cropseedc_deficit, to storvegc as well as totveg.
Send livestemc to litter for prognostic crop. Ignore downreg for carbon isotopes in Fractionation subroutine.
For dynamic roots, use rooting values on parameter dataset which is set to values from PFT=1.