clm4_5_16_r256
Changes from Leo VanKampenhout to have fresh snow grain size a function of air temperature for CLM50.
Behavior is the same as before for CLM45 with fresh snow grain size a fixed parameter.  For CLM50
it's that same value (54.526) less than -30C and linear with T up to 204.526 at 0C, and 204.526 for
temperatures greater than 0C.