clm4_5_16_r257
This is a fix for two bugs found by Hui Tang. One is for Carbon isotopes where a fix that came in for
clm4_5_11_r188, used leafc for too many of the isotope flux calculation calls. This changes it so the
correct terms are used for the particular flux.

The other fix is in the fire model where fire emissions were using the wrong fire related
mortality term for dead stem and should have been using a different type. Now it matches the
flux type, as well as the term used in combustion completness for the particular fire emission.