clm4_5_18_r268
Changes from Louisa Emmons to fix the calculation of stomatal resistence in the dry-deposition code so it's the inverse
of the previous calculation (and the arbitrary multiplicitive factor to reduce it was removed). Add new optional history
fields for dry-deposition.

Increase the threshold of total vegetation carbon for when decomposition pools are reset from 1.e-8 to 0.1. This change comes from
Dave Lawrence and Keith Oleson. The code was also modified to use better names for this rather than ccrit and ncrit which came in
clm4_5_17_r264. Add some new checking for these settings at initialization. Add some changes to the default CMIP6 output from
Keith Oleson.

Add in f05 fsurdat and one landuse.timeseries datasets. Add check in SSP test that makes sure the compset used includes MOSART as
expected.