clm4_5_18_r269
Move the CLM externals to the version in github for MOSART, RTM and
PTCLM. For PTCLM this meant updating to a version that didn't take
advantage of subversion keywords. I also updated to a version that
can properly compare to the files it keeps even with differences
of user, machine, path, or date.