Copy as many snow layers as possible in init_interp
In most cases, we only need data from the existing snow layers. But in a
few cases - in particular, the flx_abs* variables - we need data even
from non-existing snow layers in order for interpolation to be
bit-for-bit. The change here reworks snow interpolation so that, in
addition to copying the existing snow layers, we also copy as many
non-existing snow layers as will fit in the destination layer structure.

Fixes #326