Add an LII test covering the new init_interp method
It may not be totally necessary to have a system test covering this
init_interp_method: between unit tests and inline checks, we have quite
a bit of testing of this code. It's important to have at least one LII
test to make sure we're interpolating all the fields we're supposed to
interpolate, but we don't necessarily need a LII test with this mode of
operation as long as we have one with the general mode of operation.

However, this test is a good check that every point has a unique type
(because this mode of operation will fail if that's not the case); this
is something needed even for the 'general' init_interp method to work
correctly, but the 'general' mode won't catch problems in this regard.

One other reason why this test is useful is to cover the threading
directives in the set_single_match routine, since those aren't covered
by unit tests. So this test mod should be used in an LII test with
threading.