Add a usermods directory that turns on the ICE_MODEL_FRACTION field
Also, change ICE_MODEL_FRACTION to time-average rather than
instantaneous: I think this will more often do the right thing for
various history output frequencies. For example, if someone adds this to
an every-10-year hist file, they'd probably want the 10-year-average
ICE_MODEL_FRACTION, not the value at the end of the 10-year
period. Also: The thing that originally got me worried about having it
instantaneous was not being completely sure whether the value written at
the end of the year would apply to the current year or the previous
year. I think it would apply to the previous year (which I think is what
you'd want), but I felt it's safer to just write the time-averaged
value (where this year boundary problem would, at worst, lead to errors
of 1/17520 for a 30-min time step, rather than leading to a complete
year offset -- i.e., if we average just the year-boundary value into the
wrong year).

Fixes #345