Revert "Point to new initial conditions file"
This reverts commit 21ca821fe65e5654e940f3cee53c798a24de26b6.
This reverts commit 7eaa12f43804e00ee35c7a04ee0f699c57838728.

There were answer changes for the new initial conditions file. One
problem with the new finidat file is that it doesn't have C isotope
information on it... that's a problem, but not the (only) problem with
these baseline failures, I think. Another difference is that two
variables have changed from double to int. I'm not sure if these
completely explain the problem, but I'm going to back out the change to
finidat, just using the new file for the test that needs it.