Remove the water_bulk_type and water_tracer_type containers
This is design (D5) in the pdf in PR #395, which is the favorite of
Mariana Vertenstein and Mat Rothstein.

In discussion with Mariana and Mat, we felt that the benefits of having
these containers weren't worth the extra complexity of having this extra
layer. Part of my original motivation for having this extra layer was
thinking that we'd eventually have a lot more objects contained in each
tracer instance: all of the process-oriented objects like irrigation,
infiltration excess runoff, etc. (which each compute one or more water
fluxes). But as I noted in
https://github.com/ESCOMP/ctsm/pull/395#issuecomment-392299340, I'm now
feeling that won't be feasible, so we'll probably have no more than 5 or
6 objects for each tracer instance (state, diagnostics, water balance,
fluxes, and maybe we'll split one or two of those further). So there is
less of a need for this tracer type container.