Clean up the tracer consistency check
Previously, it was both doing an endrun and returning a logical (which
would could never be false, because if it was, then the endrun would
have been called). Now we just do an endrun.

I prefer this method because it is less error-prone: If we don't do the
endrun internally but instead rely on passing a logical up the call
stack, then a single instance of accidentally having

wiso_inconsistency = WisoCompareBulkToTracer(...)

instead of

wiso_inconsistency = wiso_inconsistency .or. WisoCompareBulkToTracer(...)

will cause checks to pass when they shouldn't - and this could be really
hard to track down.

Note that the old water isotope branch had a logical controlling whether
to die internally or pass the logical up the call stack, but I feel
that's too much complexity without much benefit.