Add infrastructure to loop over all tracers
I didn't like that you needed to remember to add a call to
CompareBulkToTracer for every new variable: it would be easy to forget
this, it would be hard to catch that in a review, and tests wouldn't
catch that omission. This commit adds infrastructure that allows looping
over all water tracer variables for this or other purposes.