Maintain lower bounds of data arrays
In order to maintain the lower bounds of the data arrays, we need to
either (a) declare 'data' to be a pointer, or (b) pass in the lower
bounds as an argument. For now, since all of our actual variables are
pointers, I'm going with (a). In the future, if we change some variables
to be target, non-pointers, then we should have the caller pass in the
lower bound and change 'data' to:

    real(r8), target, intent(in) :: data(lbound:)