Revert "Maintain lower bounds of data arrays"
This reverts commit 2bf5536128911fa984becb40addd1657c1bf6d4f.

With nag, lower bounds weren't maintained for slices of 2d arrays. I
think that, if I wanted to maintain the lower bounds of those, I would
need to create a temporary pointer in AllocateVar2d that has the correct
lower bounds. It seems like it will be less error-prone to just pass the
lower bounds in. So I'm going to rework the code to work that way.