Use init method rather than constructor for water_tracer_container_type
With the old version, both system tests and unit tests the unit tests
were failing on cheyenne with intel17.0.1, and system tests were failing
on hobart with intel18.0.3.

I looked extensively into the unit test failures. The problem seemed to
be that the vector size turns into garbage in the return from the
constructor: When the test has:

  container = new_water_tracer_container_type()

We get:

 end of clear_vec: vec_size =            0
 end of new_water_tracer_container_type: size =            0
 in test after constructor: size =           97

The problem seems to come in this line:

    container = water_tracer_container_type()

I think this is invoking the copy method, vector_assign_vec, for the
vector component of the derived type.

From inserting some extra prints in this, I see:

  print *,	'in vector_assign_vec: self, other size = ', self%vsize(), other%vsize()
  allocate(temp, source=other)
  print *,	'after allocate: temp size = ',	temp%vsize()
  call self%swap(temp)
  print *,	'after swap: self size = ', self%vsize()

 printing in new_water_tracer_container_type
 in clear_vec
 in resize_vec with new_size, capacity =            0           0
 end of clear_vec: vec_size =            0
 in vector_assign_vec: self, other size =            0           0
 after allocate: temp size =            0
 after swap: self size =            0
 end of new_water_tracer_container_type: size =            0
 in vector_assign_vec: self, other size =            0           0
 after allocate: temp size =           97
 after swap: self size =           97
 in test after constructor: size =           97

Note that the first entry into vector_assign_vec is this line in
new_water_tracer_container_type:

    this%tracer_vec = water_tracer_vector()

whereas the second is this line in the test routine:

    container = water_tracer_container_type()

The first one seems fine, but the second one gives problems.

Changing the water_tracer_container_type constructor to a call to an
init method (thus avoiding the assignment) seems to resolve the problem.