When initializing new columns, also initialize water tracers
When we initialize new columns with dynamic landunits/columns,
initialize water tracer variables the same way that we initialize bulk
water variables.

I considered initializing the bulk, then initializing tracers as having
a fixed ratio times the bulk. But I think the current
implementation (where we copy the tracer state from the template column)
is better in cases where the state has evolved away from the fixed
ratio: it's more like what would happen if we were doing rigorous state
updates.

Introducing a unit test for this new behavior required a fair amount of
refactoring of existing unit tests; that is also done in this commit.