Call BeginWaterBalance for each water tracer as well as bulk
I have also prepared the BalanceCheck routine to handle multiple
tracers, but it is not yet called for each tracer. Note that it accepts
an argument of waterdiagnosticbulk_type, NOT the tracer-specific
waterdiagnostic_type: this is intentional, because we need the bulk-only
values for some snow diagnostic terms. This preparation required moving
a number of variables into types that exist for tracers as well as
bulk. I removed diagnostic prints of total_plant_stored_h2o_col and
qflx_rootsoi_col in BalanceCheck: These terms don't appear in any
equations in this routine, and they made the argument passing awkward
(total_plant_stored_h2o_col was the only term needed from the
tracer-specific waterdiagnostic_type; qflx_rootsoi_col is not currently
separated by tracer in waterflux_type - though the latter will probably
need to be changed eventually). Similarly, I removed a diagnostic print
of qflx_top_soil, which also didn't appear in any equations in
BalanceCheck.