Add ability to loop through bulk and tracers all at once
I was ending up with more instances than I had originally imagined where
I needed to do the same thing on the bulk and all tracers. In order to
remove duplication, this commit introduces a method for looping through
both bulk water and all tracers in a single loop.

One of the big motivations here is that I'm finding that there are a lot
of places in the code that need to loop over bulk and all tracers. I'm
starting to feel that there are more downsides than upsides with
requiring all of these loops to happen in WaterType.F90. So I'm thinking
of allowing other places to loop over all tracers, or bulk and all
tracers (as is now done in lnd2atmMod.F90). In order to support that, I
want this looping to be as simple as possible, including using named
bounds (rather than, for example, assuming that you should loop
0..num_tracers in order to loop over bulk and all tracers) - so that we
can change things internally in WaterType.F90 later without requiring
modifying a lot of code.