Move some loops over tracers out of WaterType into better homes
It was starting to get too awkward to require all loops over tracers to
happen in WaterType (or some other centralized location): I had
originally imagined that we wouldn't need too many loops over tracers,
but it turns out that we need loops over tracers in a lot of
places. Requiring all of these loops over tracers to be in WaterType
both (a) bloats that module, and (b) adds extra indirection (which makes
it harder to understand the code, because you're bouncing back and forth
between more modules, and has possible performance implications as we
break routines into tiny pieces for this purpose).

So this commit moves most of these loops over tracers out of
WaterType. This keeps operations all together in one module.