Only allow CMIP6DECK modifier with CLM50%BGC-CROP
Per suggestion of @ekluzek, we now use CLM50%BGC-CROP-CMIP6DECK rather
than CLM50%BGC-CROP%CMIP6DECK. One main purpose of this is so that we
can keep modifier_mode="1" (since changing it to "+" would allow many
things that are non-sensical). Also, as Erik pointed out, CMIP6DECK can
only be used with BGC-CROP; the previous implementation allowed it with
other options as well. This new version also only allows CMIP6DECK with
CLM50 - not with CLM45.

A few notes about the implementation:

- I didn't need to add anything in CLM_BLDNML_OPTS, because
  %BGC-CROP-CMIP6DECK still matches the existing "BGC-CROP" regular
  expression.

- The match for CLM_USER_MODS is something fairly
  general ("_CLM50%[^_]*CMIP6DECK"). This seemed more future-proof than
  explicitly listing "BGC-CROP-CMIP6DECK". The description section will
  currently prevent CMIP6DECK from being specified anywhere other than
  with BGC-CROP, so for the CLM_USER_MODS I figured I might as well
  match CMIP6DECK occurring anywhere within a CLM50 modifier.