Change uppercase DEBUG variables to lowercase debug
This is needed to avoid conflicting with the DEBUG CPP token.

I also changed one instance of debug to a 'parameter'. I left others as
is because I didn't do the analysis needed to confirm that they're never
set from elsewhere (because they are public to their modules).

Fixes ESCOMP/ctsm#534