Run CTSM's Fortran unit tests as part of aux_clm testing
This way, we don't need to run these as a separate job. Note that the
implementation abuses the system test infrastructure, following what was
done in https://github.com/ESMCI/cime/pull/2826.

I tried P1x1, P1x8 and P8x1 for this test. I thought that P1x8 or P8x1
might allow faster parallel build times. However, the time was nearly
the same for all three. For now I'm sticking with P1x1. It's possible
that CISL's job-killers will kill this job sometimes if we take too many
resources on the share queue when we have only requested a single
processor (due to the parallel build); if so, we can try increasing it
to P1x8 or P8x1.