Fix getting default charge account for latest cime
The latest version of cime changed how get_charge_account works so that
it no longer falls back on get_project: instead, it requires you to pass
in a project as a fallback. So the previous code stopped finding my
default project on cheyenne. I would have had to first call get_project,
then pass in the found project to get_charge_account for it to work. On
CESM machines, it turns out that we don't pay any attention to
CHARGE_ACCOUNT anyway in cime, so using get_project better matches what
we would get in a cime/cesm run.