Suggest reusing an existing directory for unit tests
Now that the unit tests are run from a fresh directory as part of
aux_clm, it is not so important that developers do that when running
them by hand. Doing an incremental rebuild greatly speeds up the unit
testing, so I now suggest doing it that way in the instructions.