Add new IC files for clm*_cam6.0 tuning
New IC files for when coupled to CAM, both 1850 and 2000/2010.
Also add in 2010 sim_year and 2010_control. Be explicit about
the tuning mode for 2000 cases and use the old 2000 period file.

When use_init_interp is NOT set in the namelist, let the default
be selected from the default file. This allows it to be turned on
when there are files that match close enough that you can with
use_init_interp=T.

Update documentation on ssp-rcp option.