Now calling collapse_crop_types for almost all cases
New goal to not limit and rather to call collapse_crop_types for as many
cases as possible. From subr. dyncrop_interp the call occurs for every
case because dyncrop_interp is not called if use_fates; however, in
subr. surfrd_veg_all we still need if (.not. use_fates) around the call.
Cheyenne_gnu tests PASS.
Cheyenne_intel PENDing.