Revert "Make "single-point" tests actually include multiple points"
This reverts commit 3a43fbd3ed4e82f686cb541754055453501af2ee.

I've decided that this adds complexity without providing huge
benefit (it gives more of a false confidence than any huge benefit). I'm
planning to go with a strategy of having as much code as possible
covered by ERP system tests, and not trying to rely on unit tests to
test multi-point behavior.