Merge together the pre and post-downscaling atm2lnd tracer updates
In order to keep the driver simpler, I want to reduce the number of
separate calls to do tracer updates. Since the non-downscaled tracer
forcings aren't needed until after the downscaled atm2lnd quantities are
computed, we can compute both non-downscaled and downscaled atm2lnd
tracer forcings after downscale_forcings. (If it ever happens that we
need the non-downscaled, gridcell-level tracer forcings earlier in the
driver loop, we can split this apart again.)

Also remove the early calls to TracerConsistencyCheck. Now that we only
set the non-downscaled water tracer forcings after downscale_forcings,
earlier calls to TracerConsistencyCheck are not expected to pass. I
wanted to reduce the calls to TracerConsistencyCheck in order to clean
up the driver loop anyway, but note that this means that people will NOT
be able to introduce earlier calls to TracerConsistencyCheck for
debugging purposes (if doing so is desired, we'll need to re-split the
pre and post-downscaling atm2lnd tracer updates).