Handle roundoff errors in the rain-to-snow conversion
Without this change, I was getting errors like this:

   ERROR in CompareBulkToTracer: tracer does not agree with bulk water
   Called from: after downscale_forcings
   Variable: forc_rain_downscaled_col
   First difference at index: 215
   Bulk  :   0.00000000000000000E+00
   Tracer:   0.66174449004242214E-23
    ENDRUN:
    ERROR: CompareBulkToTracer: tracer does not agree with bulk water

This change should fix that problem, by setting the tracer to exactly 0
when it ends up very near 0 due to roundoff issues.

Note that I have added a unit test for rain going to exactly 0; I have
not added a unit test for snow going to exactly 0 because it would not
cover any additional code (since the code for these updates is now in a
function that is shared between rain and snow).