Change wa_col cold start initialization when use_aquifer_layer is false
In the groundwater_irrigation branch
(https://github.com/ESCOMP/ctsm/pull/523), @swensosc has stopped
resetting wa_col each time step if use_aquifer_layer is false. However,
this leads to having a substantially different value of wa_col when
use_aquifer_layer is false: previously, it was reset to
aquifer_water_baseline each time step, but with the changes from
@swensosc, it stays close to its initial values, which have been 4000 in
most places. This commit changes the initial values to match the value
it was being reset to, so it simply starts at aquifer_water_baseline -
so the every-time-step-resetting to aquifer_water_baseline can be
removed without massively changing the value of wa_col.

In addition to changing the cold start initialization of wa_col, we are
also changing the cold start initialization of zwt in this case where
use_aquifer_layer is false: The old initialization of zwt wouldn't work
as intended now that we have changed the initial value of wa_col;
@swensosc suggested this new initialization method instead. This
initialization to zi(c,nbedrock(c)) is correct if there are no saturated
layers, and close enough for a decent cold start even if there are.